@@ -1,4 +1,4 @@
-class EntryIssue < ActiveRecord::Base
+class EntryIssue < ApplicationRecord
   belongs_to :entry
   belongs_to :issue, touch: true
 end
