@@ -1,4 +1,4 @@
-class EntryKeyword < ActiveRecord::Base
+class EntryKeyword < ApplicationRecord
   belongs_to :entry
   belongs_to :keyword
 end
