@@ -1,4 +1,4 @@
-class EntryTag < ActiveRecord::Base
+class EntryTag < ApplicationRecord
   belongs_to :entry
   belongs_to :tag
 end
