@@ -1,4 +1,4 @@
-class FeedTopic < ActiveRecord::Base
+class FeedTopic < ApplicationRecord
   belongs_to :feed
   belongs_to :topic, touch: true
 end
