@@ -1,4 +1,4 @@
-class Preference < ActiveRecord::Base
+class Preference < ApplicationRecord
   belongs_to :user
 
   DELETE_VALUE = "==DELETE=="
