@@ -1,4 +1,4 @@
-class SubscriptionCategory < ActiveRecord::Base
+class SubscriptionCategory < ApplicationRecord
   belongs_to :subscription
   belongs_to :category
 end
