@@ -1,4 +1,4 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
 require 'rails/all'
 
@@ -21,10 +21,6 @@ module SpreadBeaver
     # config.i18n.default_locale = :de
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
     config.api_only = false
-    config.generators do |g|
-      g.template_engine :slim
-    end
   end
 end
